+++ crates/nostr-sqlite/src/lib.rs
@@ -320,15 +320,6 @@ impl NostrDatabase for SQLiteDatabase {
     }
 
     #[inline]
-    async fn event_ids_by_filters(
-        &self,
-        filters: Vec<Filter>,
-        order: Order,
-    ) -> Result<Vec<EventId>, Self::Err> {
-        Ok(self.indexes.query(filters, order).await)
-    }
-
-    #[inline]
     async fn negentropy_items(
         &self,
         filter: Filter,