+++ CHANGELOG.md
@@ -82,6 +82,7 @@
 * nostr: remove support for NIP-44 v1 ([Yuki Kishimoto])
 * nostr: remove `EventBuilder::encrypted_direct_msg` ([Yuki Kishimoto])
 * database: remove `TempEvent` ([Yuki Kishimoto])
+* database: remove `NostrDatabase::event_ids_by_filters` ([Yuki Kishimoto])
 * sdk: remove `Client::send_direct_msg` ([Yuki Kishimoto])
 * cli: remove `tracing-subscriber` dep